define AsymmetricEncryptor trait with RSA implementation for better ergonomics
