+++ src/crypto/mod.rs
@@ -1,3 +1,3 @@
+pub mod asymmetric;
 pub mod errors;
-pub mod rsa;
 pub mod symmetric;
