--- src/crypto/rsa/mod.rs
@@ -1,2 +1,0 @@
-pub mod cryptor;
-pub mod keygen;
